ocaml: force -j1 when building ocamlopt
Without the ocamlopt flag, parallel build works fine, but when building
the native compiler, parallel compilation fails.
This patch allows all of the bytecode compiler to still be compiled in
parallel, and if the native compiler (ocamlopt) is built, only that will
be built with jobs=1.

This patch also adds a missing apostrophe within a comment :)